♻️ 사진 추가 util 함수 구현
